# gateway/kitty-chat/src/lib/gen/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JournalChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Messages",
            "type": "array"
          },
          "system_prompt": {
            "default": "",
            "title": "System Prompt",
            "type": "string"
          }
        },
        "required": [
          "messages"
        ],
        "title": "JournalChatRequest",
        "type": "object"
      },
      "JournalJournalChatResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "JournalJournalChatResponse",
        "type": "object"
      },
      "JournalJournalPromptResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "JournalJournalPromptResponse",
        "type": "object"
      },
      "JournalJournalStartResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "JournalJournalStartResponse",
        "type": "object"
      },
...
      "JournalSessionsSessionIdMessagesMessageIdResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "JournalSessionsSessionIdMessagesMessageIdResponse",
        "type": "object"
      },
...
      "LifeLifeDismissSignalKindResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "LifeLifeDismissSignalKindResponse",
        "type": "object"
      },
...
      "LifeLifeEventsResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "LifeLifeEventsResponse",
        "type": "object"
      },
...
      "LoopsLoopLoopIdResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "LoopsLoopLoopIdResponse",
        "type": "object"
      },
      "LoopsLoopLoopIdToggleResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "LoopsLoopLoopIdToggleResponse",
        "type": "object"
      },
      "LoopsLoopsResponse": {
        "additionalProperties": true,
        "properties": {},
        "title": "LoopsLoopsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kitty Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/journal/chat": {
      "post": {
        "description": "Single journal interview turn \u2014 uses provided system_prompt, bypasses context assembler.",
        "operationId": "journal_chat_journal_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalJournalChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Journal Chat",
        "tags": [
          "journal"
        ]
      }
    },
    "/journal/prompt": {
      "get": {
        "description": "Return a random journal writing prompt. Optional ?theme= filter.",
        "operationId": "journal_prompt_journal_prompt_get",
        "parameters": [
          {
            "in": "query",
            "name": "theme",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theme"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalJournalPromptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Journal Prompt",
        "tags": [
          "journal"
        ]
      }
    },
    "/journal/start": {
      "post": {
        "description": "Begin a journal interview session. Returns Kitty's opening question.",
        "operationId": "journal_start_journal_start_post",
        "parameters": [
          {
            "in": "query",
            "name": "theme",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theme"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalJournalStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Journal Start",
        "tags": [
          "journal"
        ]
      }
    },
...
    "/life/dismiss/{signal_kind}": {
      "post": {
        "operationId": "post_dismiss_life_dismiss__signal_kind__post",
        "parameters": [
          {
            "in": "path",
            "name": "signal_kind",
            "required": true,
            "schema": {
              "title": "Signal Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifeLifeDismissSignalKindResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Dismiss",
        "tags": [
          "life"
        ]
      }
    },
...
    "/life/events": {
      "get": {
        "operationId": "list_life_events_life_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifeLifeEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Life Events",
        "tags": [
          "life"
        ]
      }
    },
...
    "/loop/{loop_id}": {
      "delete": {
        "description": "Delete a loop.",
        "operationId": "delete_loop_loop__loop_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "loop_id",
            "required": true,
            "schema": {
              "title": "Loop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopsLoopLoopIdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Loop",
        "tags": [
          "loops"
        ]
      }
    },
    "/loop/{loop_id}/toggle": {
      "post": {
        "description": "Toggle a loop on/off.",
        "operationId": "toggle_loop_loop__loop_id__toggle_post",
        "parameters": [
          {
            "in": "path",
            "name": "loop_id",
            "required": true,
            "schema": {
              "title": "Loop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopsLoopLoopIdToggleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Loop",
        "tags": [
          "loops"
        ]
      }
    },
    "/loops": {
      "get": {
        "description": "Get all active loops from the real cron schedule store.",
        "operationId": "get_loops_loops_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopsLoopsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Loops",
        "tags": [
          "loops"
        ]
      },
      "post": {
        "description": "Create a new loop as a cron schedule.",
        "operationId": "create_loop_loops_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Loop",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopsLoopsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Loop",
        "tags": [
          "loops"
        ]
      }
    },
...
    "/sessions/{session_id}/messages/{message_id}": {
      "delete": {
        "description": "Delete a specific message from a session's journal by message_id (timestamp).",
        "operationId": "delete_message_sessions__session_id__messages__message_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalSessionsSessionIdMessagesMessageIdResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Message",
        "tags": [
          "journal"
        ]
      }
    },
...
  }
}